Just rename CheckGcpEnvironment to InternalCheckGcpEnvironment instead of extra layer of wrapping.
